feat(dotnet): AgentHostProtocol .NET client at Swift/TS parity
Full .NET client (net8.0;net9.0): codegen (generate-csharp.ts), reducers,
transport, AhpClient, MultiHostClient, WebSockets — at Swift/TS feature + test
parity. 315 tests green on both TFMs under xUnit 3 + Microsoft.Testing.Platform,
gated by a machine-checked parity manifest + count floor (CI-blocking + a local
pre-push ratchet). Production features mirrored from Swift: HostEvent.Removed,
UnsubscribeAsync (+ typed errors), explicit clientSeq override, typed host errors.
Spec-conformance fix: schema-required nested objects emit C# `required`.

Bundles the shared language-neutral round-trip fixture corpus its round-trip
tests consume, so this PR is self-contained (independently reviewable/mergeable).

+++ clients/kotlin/src/main/kotlin/com/microsoft/agenthostprotocol/Reducers.kt
@@ -197,7 +197,7 @@ private fun now(): Long = currentTimestampProvider()
 // ─── Status Bitset Helpers ──────────────────────────────────────────────────
 
 /** Bitmask covering the mutually-exclusive activity bits (bits 0–4). */
-private const val STATUS_ACTIVITY_MASK: Int = (1 shl 5) - 1
+private const val STATUS_ACTIVITY_MASK: Long = (1L shl 5) - 1L
 
 /** Sets or clears a metadata flag on a status value. */
 private fun withStatusFlag(status: SessionStatus, flag: SessionStatus, set: Boolean): SessionStatus =
